+++ R/PivotTable.R
@@ -175,6 +175,10 @@
 #'   includeGroupValues=TRUE, separator=" ")}}{Gets the pivot table as a tidy
 #'   data frame, where each cell in the body of the pivot table becomes one row
 #'   in the data frame.}
+#'   \item{\code{asBasicTable = function(exportOptions=NULL, compatibility=NULL,
+#'   showRowGroupHeaders=FALSE))}}{Generates a basictabler table (from the
+#'   basictabler R package) which allows further custom manipulation of the
+#'   pivot table.}
 #'   \item{\code{getCss(styleNamePrefix)}}{Get the CSS declarations for the
 #'   entire pivot table.}
 #'   \item{\code{getHtml(styleNamePrefix, includeHeaderValues=FALSE,
@@ -1515,16 +1519,17 @@ PivotTable <- R6::R6Class("PivotTable",
       if(private$p_traceEnabled==TRUE) self$trace("PivotTable$asTidyDataFrame", "Got pivot table as a tidy data frame.")
       return(invisible(as.data.frame(df, stringsAsFactors=stringsAsFactors)))
     },
-    asBasicTable = function(exportOptions=NULL, compatibility=NULL) {
+    asBasicTable = function(exportOptions=NULL, compatibility=NULL, showRowGroupHeaders=FALSE) {
       timeStart <- proc.time()
       if(private$p_argumentCheckMode > 0) {
         checkArgument(private$p_argumentCheckMode, TRUE, "PivotTable", "asBasicTable", exportOptions, missing(exportOptions), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list")
         checkArgument(private$p_argumentCheckMode, TRUE, "PivotTable", "asBasicTable", compatibility, missing(compatibility), allowMissing=TRUE, allowNull=TRUE, allowedClasses="list", allowedListElementClasses=c("character", "integer", "numeric", "logical"))
+        checkArgument(private$p_argumentCheckMode, TRUE, "PivotTable", "asBasicTable", showRowGroupHeaders, missing(showRowGroupHeaders), allowMissing=TRUE, allowNull=FALSE, allowedClasses="logical")
       }
       if(private$p_traceEnabled==TRUE) self$trace("PivotTable$asBasicTable", "Converting to basic table...")
       if(!private$p_evaluated) stop("PivotTable$getHtml():  Pivot table has not been evaluated.  Call evaluatePivot() to evaluate the pivot table.", call. = FALSE)
       if(is.null(private$p_cells)) stop("PivotTable$getHtml():  No cells exist to render.", call. = FALSE)
-      btbl <- convertPvtTblToBasicTbl(self, exportOptions, compatibility)
+      btbl <- convertPvtTblToBasicTbl(self, exportOptions, compatibility, showRowGroupHeaders=showRowGroupHeaders)
       if(private$p_traceEnabled==TRUE) self$trace("PivotTable$asBasicTable", "Converted to basic table.")
       private$addTiming("asBasicTable", timeStart)
       return(invisible(btbl))